fix(iot-dev): Fix bug where AMQP layer would block a while on closing if the proton reactor thread crashed
still need to do manual testing
