+++ device/iot-device-client/src/main/java/com/microsoft/azure/sdk/iot/device/transport/amqps/ReactorRunnerStateCallback.java
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.microsoft.azure.sdk.iot.device.transport.amqps;
+
+interface ReactorRunnerStateCallback
+{
+    void onReactorClosedUnexpected();
+}
